+++ src/cli/commands.ts
@@ -5,35 +5,20 @@
  * Copyright (c) 2023 Yan
  */
 
-import fs from "node:fs";
-import path from "node:path";
-import * as plugins from "./modules";
-
-function detectModule(url?: string) {
-  if (!url) return undefined;
-  const downloaders = Object.values(plugins);
-  for (let j = 0; j < downloaders.length; j++) {
-    const Downloader = downloaders[j];
-    if (Downloader.canHandleUrl(url)) {
-      return Downloader;
-    }
-  }
-}
-
-function findModule(name: string) {
-  const downloaders = Object.values(plugins);
-  for (let j = 0; j < downloaders.length; j++) {
-    const Downloader = downloaders[j];
-    if (Downloader.siteName === name) {
-      return Downloader;
-    }
-  }
-}
+import fs from 'node:fs';
+import path from 'node:path';
+import {
+  detectModule,
+  findModule,
+  genModule,
+  genPresets,
+  mergeOptions,
+} from './lib';
 
 function buildDownloader(options: Partial<CliOptions> = {}) {
   const {
     module,
-    output,
+    output = '.',
     cookie,
     archive,
     timeout,
@@ -44,14 +29,18 @@ function buildDownloader(options: Partial<CliOptions> = {}) {
     zipLevel,
     url,
     format,
+    auth,
   } = options;
   const Downloader = module?.length ? findModule(module) : detectModule(url);
   if (!Downloader) {
-    throw new Error("Module not found.");
+    throw new Error('Module not found.');
+  }
+  if (!silence || verbose) {
+    console.log(`Using ${Downloader.siteName}.`);
   }
 
-  const downloader = new Downloader(output ?? ".", {
-    cookie: cookie && fs.readFileSync(path.resolve(cookie)).toString(),
+  const downloader = new Downloader(output, {
+    cookie: cookie ? fs.readFileSync(path.resolve(cookie)).toString() : auth,
     timeout,
     silence,
     batchSize: batch ?? 1,
@@ -65,7 +54,8 @@ function buildDownloader(options: Partial<CliOptions> = {}) {
   return downloader;
 }
 
-export const listCommand: Command = async (name, sub, options = {}) => {
+export const listCommand: Command = async (name, sub, _options = {}) => {
+  const options = mergeOptions(_options);
   const { url, verbose, silence, output, name: rename } = options;
 
   try {
@@ -75,15 +65,15 @@ export const listCommand: Command = async (name, sub, options = {}) => {
       if (serie) {
         if (!silence) {
           console.log(`Title: ${serie.title}`);
-          console.log("----");
-          serie.chapters?.forEach((e) => {
+          console.log('----');
+          serie.chapters?.forEach(e => {
             console.log(`${e.index}  ${e.name}`);
             console.log(`${e.uri}`);
-            console.log("----");
+            console.log('----');
           });
 
           if (serie?.info) {
-            Object.keys(serie.info).forEach((e) => {
+            Object.keys(serie.info).forEach(e => {
               console.log(`${e}: ${serie.info?.[e]}`);
             });
           }
@@ -93,7 +83,7 @@ export const listCommand: Command = async (name, sub, options = {}) => {
           await downloader.writeComicInfo(serie, { output, rename });
         }
       } else {
-        console.log("Please Provide URL.");
+        console.log('Please Provide URL.');
       }
     }
   } catch (error) {
@@ -101,13 +91,14 @@ export const listCommand: Command = async (name, sub, options = {}) => {
       console.error(error);
     } else {
       console.log(
-        "Failed to get serie info. (Use -v or --verbose flag for detailed error messages.)"
+        'Failed to get serie info. (Use -v or --verbose flag for detailed error messages.)',
       );
     }
   }
 };
 
-export const downloadCommand: Command = async (name, sub, options = {}) => {
+export const downloadCommand: Command = async (name, sub, _options = {}) => {
+  const options = mergeOptions(_options);
   const {
     url,
     from,
@@ -121,7 +112,7 @@ export const downloadCommand: Command = async (name, sub, options = {}) => {
     override,
   } = options;
 
-  let current: Partial<DownloadProgress> = {};
+  const current: Partial<DownloadProgress> = {};
 
   try {
     if (url) {
@@ -136,40 +127,41 @@ export const downloadCommand: Command = async (name, sub, options = {}) => {
         override,
         chapters:
           chapters !== undefined
-            ? `${chapters}`.split(",").map((e) => parseInt(e))
+            ? `${chapters}`.split(',').map(e => parseInt(e))
             : undefined,
       });
     } else {
-      console.log("Please Provide URL.");
+      console.log('Please Provide URL.');
     }
   } catch (error) {
     if (verbose) {
       console.error(error);
     } else {
       console.log(
-        "Download Failed. (Use -v or --verbose flag for detailed error messages.)"
+        'Download Failed. (Use -v or --verbose flag for detailed error messages.)',
       );
     }
 
     if (current.index) {
       console.log(`Latest download: [${current.index}] ${current.name}.`);
       console.log(
-        `Status: ${current.status}, Pages not downloaded: ${current.failed}.`
+        `Status: ${current.status}, Pages not downloaded: ${current.failed}.`,
       );
       console.log(
         `Retry from the next chapter: Use -f ${
-          current.status === "completed" ? current.index + 1 : current.index
-        } ${to ? `-t ${to}` : ""}`
+          current.status === 'completed' ? current.index + 1 : current.index
+        } ${to ? `-t ${to}` : ''}`,
       );
     } else {
       console.log(
-        "No chapter is downloaded, please check the availabiliy of the module (site) or your Internet connection."
+        'No chapter is downloaded, please check the availabiliy of the module (site) or your Internet connection.',
       );
     }
   }
 };
 
-export const chapterCommand: Command = async (name, sub, options = {}) => {
+export const chapterCommand: Command = async (name, sub, _options = {}) => {
+  const options = mergeOptions(_options);
   const { url, name: chapterName, verbose, output, override } = options;
 
   try {
@@ -179,20 +171,34 @@ export const chapterCommand: Command = async (name, sub, options = {}) => {
       if (output) {
         serie = await downloader.getSerieInfo(url);
       }
-      await downloader.downloadChapter(chapterName ?? "Untitled", url, {
+      await downloader.downloadChapter(chapterName ?? 'Untitled', url, {
         info: serie?.info,
         override,
       });
     } else {
-      console.log("Please Provide URL.");
+      console.log('Please Provide URL.');
     }
   } catch (error) {
     if (verbose) {
       console.error(error);
     } else {
       console.log(
-        `Failed to download ${chapterName}. (Use -v or --verbose flag for detailed error messages.)`
+        `Failed to download ${chapterName}. (Use -v or --verbose flag for detailed error messages.)`,
       );
     }
+  }
+};
+
+export const genCommand: Command = async (_, sub, options = {}) => {
+  const { module, presets } = options;
+
+  if (module) {
+    await genModule(module);
+  } else if (presets) {
+    await genPresets();
+  } else {
+    throw new Error(
+      'The gen command is used to generate a module or a presets file, please use the --module or --presets flag.',
+    );
   }
 };