+++ src/global.d.ts
@@ -9,7 +9,7 @@ interface Configs {
   silence?: boolean;
   batchSize?: number;
   verbose?: boolean;
-  archive?: "zip" | "cbz" | boolean;
+  archive?: 'zip' | 'cbz' | boolean;
   headers?: Record<string, any>;
   maxTitleLength?: number;
   zipLevel?: number;
@@ -20,6 +20,7 @@ interface Chapter {
   index: number;
   name: string;
   uri?: string;
+  options?: Record<string, any>; // this passes to getImageList
 }
 
 /**
@@ -41,8 +42,8 @@ interface ComicInfo {
   Tags?: string;
   Web?: string;
   Note?: string;
-  Manga?: "YesAndRightToLeft";
-  Status?: "Ongoing" | "End" | "Abandoned" | "Hiatus"; // This one isn't in the schema but it should
+  Manga?: 'YesAndRightToLeft';
+  Status?: 'Ongoing' | 'End' | 'Abandoned' | 'Hiatus'; // This one isn't in the schema but it should
   [Key: string]: any; // and many more
 }
 
@@ -65,8 +66,9 @@ interface DownloadProgress {
   index?: number;
   name: string;
   uri?: string;
-  status: "completed" | "failed" | "skipped";
+  status: 'completed' | 'failed' | 'skipped';
   failed?: number;
+  options?: Record<string, any>; // options from Chapter
 }
 
 interface SerieDownloadOptions {
@@ -109,7 +111,7 @@ interface CliOptions {
   cookie: string;
   from: number;
   to: number;
-  archive: "zip" | "cbz" | boolean;
+  archive: 'zip' | 'cbz' | boolean;
   timeout: number;
   silence: boolean;
   batch: number;
@@ -122,10 +124,13 @@ interface CliOptions {
   info: boolean;
   format: string;
   override: boolean;
+  presets: string; // path
+  auth: string;
+  [key: string]: any; // for future modules
 }
 
 type Command = (
   name: string,
   sub: string[],
-  options?: Partial<CliOptions>
+  options?: Partial<CliOptions>,
 ) => void;