+++ src/modules/index.ts
@@ -1,2 +1,3 @@
 export { default as ZeroBywDownloader } from "./zerobyw";
 export { default as CopymangaDownloader } from "./copymanga";
+export { default as GanmaDownloader } from "./ganma";